Remove unnecessary PostTask from SafeBrowsingQueryManager
Right now the SafeBrowsingQueryManager always asynchronously runs the
callback. This could happen synchronously, with asynchronous work
happening on demand. This requires a lot of test changes because the
FakeSafeBrowsingService does not model the real SafeBrowsingUrlCheckImpl
very accurately.

The real SafeBrowsingUrlCheckImpl calls
`url_checker_delegate_->StartDisplayingBlockingPageHelper` when an
unsafe URL is found. It then waits for the callback on the
UnsafeResource before proceeding. The FakeSafeBrowsingService currently
relies on the async nature of the SafeBrowsingQueryManager to let tests
manually perform the same work StartDisplayingBlockingPageHelper would
do.

This is easily fixed by making the FakeSafeBrowsingService more accurate
to the real implementation and updating all the relevant tests.

Bug: 359420122
Change-Id: I0a8b755a20b00ddc080978bf1509be49c1b5b2b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5852215
Reviewed-by: Joemer Ramos <joemerramos@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1361960}