Remove obsolete comment in DelegatedInkTrailPresenter
Clips and transforms to the presenter area are now taken into
consideration. The delegated ink trail will follow the pointer
regardless of the transformation to the presenter area. Additionally,
the ink trail will not be drawn outside of the clip boundary so long as
the client does not continue to send updateInkTrailStartPoint calls.
Prior to CL:5801881 the ink trail would persist.

Bug: 345493182
Change-Id: I2b23a9e8fd52be4afb51befe8c3130bb10231038
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5900493
Commit-Queue: Sahir Vellani <sahir.vellani@microsoft.com>
Reviewed-by: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1361974}