PA KVv2: Implement compression group parsing.
For the Protected Audience KVv2 trusted signals API, the caching
implementation passes data to the Javascript process for entire
compression groups at a time. This CL adds a method to completely parse
a single compression group to trusted_signals_kvv2_helper.cc.

Unlike other methods, this API does not currently take a list of
interest groups or keys to extract, so that parsed data can be used
to handle requests that are received after the process has already
received and parsed the response body (we may try and improve this
later).

Bug: 333445540
Change-Id: I5b8fc1687ccda382dba63d95cbb057072558eb00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5883079
Reviewed-by: Russ Hamilton <behamilton@google.com>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1361976}