firmware_update: add needs_reboot field to device
This CL is part of a small chain that will conditionally change the
update complete dialog [1] of the fwupd UI if the device needs to reboot
to install an update. This is necessary for go/reven-firmware-update-dd,
where firmware updates installed via capsule may need to reboot one or
more times to install. To determine this, we look for the
`FWUPD_DEVICE_FLAG_NEEDS_REBOOT` flag [2] when querying fwupd for
devices and add that to the FwupdDevice struct. This will then attached
to an update by the update manager [3] then used in the UI accordingly.

[1]: http://cs/h/chromium/chromium/src/+/main:ash/webui/firmware_update_ui/resources/firmware_update_dialog.ts?l=339-345
[2]: https://github.com/fwupd/fwupd/blob/main/libfwupd/fwupd-enums.h#L304-L311
[3]: http://cs/h/chromium/chromium/src/+/main:chromeos/ash/components/fwupd/firmware_update_manager.cc

Test: chromeos_unittests --gtest_filter=FwupdClient*
Bug: b:361395262
Change-Id: If1d7c475e221e0946abf73aef34f316beae0640b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5888326
Commit-Queue: Andre Braga <andrebraga@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1361990}