Fix MP2T fuzzer issue
Looks like the extra_bits is zero and causing the string to resize to
nothing triggering an assert in string::front. There's a DCHECK for this
in the BitReader class, but we need an explicit check here too.

R=dalecurtis

Fixed: 362815109
Change-Id: I376a1c5190def230b5a544c92469ea67a59fbde0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5894234
Auto-Submit: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362003}