Replace generated top-level C array with base::span
TLDR: instead of generating a:
- extern const TYPE kArrayName[];
- extern const size_T kArrayNameLength;
the code should now generate:
- const base::span<const TYPE> kArrayName;

This change ensures the json-to-struct exports a base::span when a
'generate_array' rule is invoked. Migration to generate_array allows
us to resolve multiple unsafe buffer access warnings.

This is the simplest change that gets the job done. It does not
improve the health of the json to struct in any way.

Bug: b/40285824
Change-Id: I0761393c47667749f0df0389b58e1b7cf47fa485
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5894206
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1362032}