[Tab Switcher] Introduce TabGroupColorViewProvider
TabGroupColorViewProvider will be used to supply a view containing a
tab group color for the tab switcher in both LIST and GRID mode.
Previously, there were bespoke implementations for both as LIST mode
used a standalone ImageView while GRID re-used the existing favicon
infrastructure.

In the near future the TabGroupColorViewProvider will be expanded to
also supply a SharedImageTiles sub-view if the group is a
collaboration.

This CL also introduces TabCardViewBinderUtils which can hold common
static methods used in at least two of Strip, List, and Grid view
binders.

In an effort to keep CLs small this CL is a precursor to landing:
https://chromium-review.googlesource.com/c/chromium/src/+/5894504
which adds all the wiring to actually use this provider in favor of
the previous systems.

Bug: 362282115
Change-Id: I1328fe310cb5fb558eff8768e7566eda48df0901
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5898654
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362075}