Reland 1: Reduce fragility in SidePanelCoordinator (part 13 / N)
The original code had expectation logic in the setup code for all tests
that verified the registered entries, with additional logic in
AddTabToBrowser() to remove automatically added entries. The original CL
removed the logic in AddTabToBrowser(), which had different effects on
different platforms and build configurations.

This CL simply deletes the expectations. The expectations are not what
any of the tests are actually trying to verify, and instead become a
change-detector that simply validates whatever the default behavior is
on each platform, which is not useful.

> Reduce fragility in SidePanelCoordinator (part 13 / N)
>
> Previously, SidePanelCoordinator unit tests inherited from
> TestWithBrowserView. This is an anti-pattern as these tests are
> equivalent to browser_tests. This CL makes the tests browser_tests.
>
> This is primarily a refactor with no intended functionality change. Some
> tests were enabled as animations do not work on macOS unit_tests. The
> SidePanelPinButtonsHideInGuestMode test was modified to actually open a
> guest window.
>
> Bug: 363743081, 360163254
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5889112
> Reviewed-by: Caroline Rising <corising@chromium.org>
> Commit-Queue: Erik Chen <erikchen@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1360780}

Change-Id: I8f694871fa6d2f315f7ec1f5122234665c3f1850
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5894041
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362098}