[Settings] Avoid Instrumentation.startActivitySync
We're soon enabling the single-activity mode for the settings UI where a
single activity instance is used throughout the settings hierarchy. To
avoid launching a new activity while we continue to use intents for user
navigation, we set FLAG_ACTIVITY_SINGLE_TOP to intents if the
single-activity mode is enabled.

Currently site settings tests use SiteSettingsTestUtil that calls
Instrumentation.startActivitySync to start the settings activity, but
unfortunately it hangs for an intent with FLAG_ACTIVITY_SINGLE_TOP
because it does not start a new activity.

This patch changes SiteSettingsTestUtil to use Context.startActivity()
instead, combined together with
ApplicationTestUtils.waitForActivityClass to wait for the activity to be
resumed.

Bug: b/356743945
Change-Id: I84731d8599872b1729f6779c358173449f4fb9b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5874682
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362142}