Prerender: Add non-negative check to prerender PLMO
We're going to build paint tree during prerendering, which is setting
painting related timing and we should update this kind of metrics
during prerendering.

This CL adds the non-negative check for fcp and lcp, to ensure
a prerendering page does not set painting timestamp.

Bug: 40240492
Change-Id: Ifd788c4b04f06a6b98caff1bb9df55af5fb8f07a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5890664
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362154}