HttpStreamPool: Don't process pending requests on shutting down
An HttpNetworkSession owns an HttpStreamPool, a SpdySessionPool, and
an HttpNetworkSessionContext. When destructing the HttpNetworkSession,
a StreamSocket used by a SpdySession is returned to HttpStreamPool.
When a StreamSocket is returned, the HttpStreamPool tries to process
pending requests (if any). However, at the time the StreamSocket is
released, the HttpNetworkSessionContext is already destroyed due to
the order of destruction in HttpNetworkSession fields. This could
result in accessing destroyed HttpNetworkSessionContext in
HttpNetworkSession::ShouldForceQuic().

To avoid the situation, add a flag to indicate that the
HttpNetworkSession is shutting down. When the flag is true,
the HttpStreamPool doesn't process pending requests.

Bug: 369744951
Change-Id: Ia75a021d78781a2c442c2b42c1161f29517694d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5891626
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362179}