Set navbar divider color to match navbar color when drawing toEdge
Currently, the divider color can keep an opaque holdover color even as the navbar is set to transparent when drawing toEdge. For correctness, and to avoid issues on certain devices, the divider color should also be set to transparent when the navbar is.

Additionally, when drawing a transparent navbar, contrast enforcement from the window should be turned off to avoid triggering a scrim.

Bug: 41487034
Change-Id: I53420acb60749f5355e3bff884e2069ff84ac6d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5900453
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Charles Hager <clhager@google.com>
Cr-Commit-Position: refs/heads/main@{#1362192}