[Blink] Make webrtc_video_encoder plane size computations precise
The code currently computes plane sizes as stride * num_rows. However,
the last row is not guaranteed to have padding after it, so this is
technically unsafe. This CL changes the code to compute the plane size
precisely as stride * (num_rows - 1) + row_size for precision. This is
in preparation for an upcoming change to have this code use the
spanified version of ScopedMapping::Memory(), which does this same size
computation internally.

Bug: 369213516
Change-Id: I83fcd6addb66605e390d8681f9f8c18c59cf12c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5893775
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362223}