[Android] Create PlusAddressCreationErrorStateInfo struct.
This CL moves the logic for error state strings from the plus
address creation view to the controller. This is done by creating
a generic PlusAddressCreationErrorStateInfo struct to store the
UI strings info. This way, the view has a generic method to show
an error state. Thus, the logic in the view is minimized.

Bug: 354881207
Change-Id: I702b3ff2b2d1d8bda7c7b6843757b1f2db151c3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5899433
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Reviewed-by: Friedrich Hauser <friedrichh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362256}