[IdCheck] Use GetBiometricAvailabilityStatus for payments
Updates AutofillPaymentMethodsFragment to use the new method in place of
old one. Context: The intention is to remove
canUseAuthenticationWithBiometric and canUseAuthenticationWithBiometric-
OrScreenLock methods and replace them with
getBiometricAvailabilityStatus. This new API may return 4 values:
- REQUIRED - biometric is mandatory,
- BIOMETRICS_AVAILABLE - biometric auth is available but not mandatory,
- ONLY_LSKF_AVAILABLE - auth with pin or patter is available,
- UNAVAILABLE - no authentication method is available.

REQUIRED is the new value introduced to support the new Android feature
mandatory biometric re-auth (it's meant to trigger biometric auth always
when the user is in a non-trusted location). BIOMETRICS_AVAILABLE stands
for canUseAuthenticationWithBiometric and ONLY_LSKF_AVAILABLE represents
the case when no biometric but only screen lock auth is available.

Bug: 367683580
Change-Id: Icf624afc3252c3fe7b9b2a817ee379c4fbde8ce6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5893629
Reviewed-by: Vinny Persky <vinnypersky@google.com>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1362286}