Split IPP code into fine-grained targets
This provides tighter and clearer control over dependencies and makes
the build graph more efficient due to improved parallelism. More
importantly, this ensures that a given build only pulls the code that
it actually requires. This matters because IPP code is pulled in a
variety of heterogenous build environments, and not all IPP code is
buildable in all environments. For example, it does not make sense for
Chrome/Clank builds to pull in android/, some build configs do not
allow pulling in Mojo, etc.

Bug: b:352005196
Change-Id: Id5fee5769518a0997b9a3f1faacb3db1586a3353
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5890826
Commit-Queue: Etienne Dechamps <edechamps@google.com>
Reviewed-by: Rupert Wiser <bewise@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Dustin Mitchell <djmitche@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362295}