[High5] PIN Setup screen - Clear data explicitly
Instead of indirectly clearing the 'AuthData' when the screen is hidden,
explicitly clear it before invoking the exit callback. This makes it
clearer when this step is happening. This is preparing the screen to be
used for different setup methods in the future, either as a main factor,
or as an additional factor. Currently, the PIN setup screen is always the last step in the flow, but this will not be the case in the future.

Bug: b/365059362, b/348326316
Change-Id: Id42ec61c7bd2a2019cbd7b22c07451a8a4d08b3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5899442
Commit-Queue: Renato Silva <rrsilva@google.com>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362397}