Compute attribution headers rather than mutating the URLRequest
This is a cosmetic change for now, but makes it easier to store the
modifications and apply them to future URLRequests that correspond to
the same ResourceRequest.

This has a side benefit of making the function more easily testable.

Bug: 366284840
Change-Id: I27297fcae35743cc26918b6ceac4898ba9588b66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5900752
Auto-Submit: Chris Fredrickson <cfredric@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362398}