[iOS] Format passkey creation date in CPE
This CL formats the creation date of passkeys to be shown in the CPE UI using the iOS NSDateFormatter. Since the format currently found for passkeys in the Password Manager can hardly be matched using the system's formatter, it was updated to match that of the CPE.

Before/after on both surfaces:
https://screenshot.googleplex.com/74Qy94ZezStohJL.png

Low-Coverage-Reason: TESTS_IN_SEPARATE_CL Added code is covered with tests. More tests will need to be added in order for credential_details_view_controller.mm to reach the coverage requirement.
Fixed: 364321808
Change-Id: I3111c6b1e21cda6fc7bbc2dda9dc9dc720b25ca2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5886841
Commit-Queue: Noémie St-Onge <noemies@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362427}