Add metric to log whether familiar site was previously unfamiliar
This CL adds FAMILIAR_LIKELY_PREVIOUSLY_UNFAMILIAR
to SafeBrowsing.SiteProtection.FamiliarityHeuristic to record when a
site is visited for which:
- SiteFamiliarityHeuristicName::kVisitedMoreThanADayAgo applies
AND
- SiteFamiliarityHeuristicName::kGlobalAllowlistMatch does not apply
AND
- SiteFamiliarityHeuristicName::kVisitedMoreThanADayAgo did not apply
  during the previous visit
AND
- SiteFamiliarityHeuristicName::kNoVisitsToAnySiteMoreThanADayAgo did
  not apply during the previous visit

The new metric value assumes that the high confidence allowlist is
stable and that if a site currently matches the high confidence
allowlist that it probably would have matched the high confidence
allowlist at the time of the previous visit.
The new metric ignores site engagement score because site engagement
score is capped for engagement which all occurred on the same day.

BUG=369756823

Change-Id: I957db17339acf0f9578321d42859706ce59ac169
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5891675
Commit-Queue: Peter Kotwicz <pkotwicz@chromium.org>
Reviewed-by: Javier Castro <jacastro@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362524}