Sunfish: Add action button API
Add a basic API to add new action buttons for a capture mode
session. Currently, the API takes a callback, text, and an image,
and makes a PillButton from them. We will likely also need some
sort of ranking criterion in the future. The API should not do
anything if we are not in an active sunfish session, and/or there
is no region selected.

Bug: b/366621847
Test: SunfishTest.AddActionButton
Change-Id: Ic31a82a644d648eef3bf834d6c91fae9038c86f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5882469
Commit-Queue: Elijah Hewer <hewer@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362533}