[a11y] Changed location changes to a new type ui::AXLocationAndScrollUpdates.
The reason for this change is to allow, in a future CL, to serialize scroll changes alongside location changes. This CL starts by introducing a new type replacing the vector of AXLocationChanges and also includes all the plumbing and renames.

The CL doesn't add any new logic at all just yet. This CL and its children are a replica of the changes here but in a more organized way: https://chromium-review.googlesource.com/c/chromium/src/+/5757305

Change-Id: I2c269cd2b3d0e0b6f0d862eb00d78f2ce8a68ca9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5872830
Commit-Queue: Ahmed Elwasefi (Ahmad45123) <a.m.elwasefi@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362554}