[infra] Use edges to mixins for variants.
Currently when a variant is defined in starlark, it records the mixins
in a property on the node. The mixins get written out directly when
generating variants.pyl. This provides no checking that the mixins exist
and so a typo wouldn't result in an error until running
generate_buildbot_json.py. This updates the variant definitions to
actually create edges to the mixins so generating the config files from
starlark would result in an error.

Bug: 40258588
Change-Id: I5a4d90de2c7347b23b1ce0d7b67318fa15dad198
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5900394
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1362562}