Fix toast shown when web contents use full screen on mac
The web contents full screen API is different from the browser window
full screen on Mac; when a web contents uses the full screen API, the
toast controller incorrectly assumes the toast shouldn't show over
the web contents and causes the toast to be clipped off screen when
shown.

This CL allows the toast controller to check whether the window or web
contents is full screen on mac to determine anchor behavior of the toast.

Fixed: 368007145
Change-Id: I8e4b84470ca419d5546123ab6a763cef5c3d977f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5892412
Commit-Queue: Steven Luong <stluong@chromium.org>
Reviewed-by: Tom Burgin <bur@google.com>
Reviewed-by: Alison Gale <agale@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1362568}